requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-comments-xtd==2.4.0
django-contrib-comments==1.9.1
django-crispy-forms==1.7.2
django-extra-views==0.12.0
django-formset-js-improved==0.5.0.2
django-jquery-js==3.1.1
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
docutils==0.14
idna==2.8
oauthlib==3.0.1
python3-openid==3.1.0
pytz==2018.9
randomcolor==0.4.4.5
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.24.1
